+++ TestAzureBlobStorageUrl.ps1
@@ -1,0 +1,9 @@
+function Test-AzureBlobStorageUrl {
+    param(
+        $Url
+    )
+
+    $pattern = "abs://(.*)\.blob\.core\.windows\.net\/(.*)\/(.*)\?(.*)$"
+
+    [regex]::Match($Url, $pattern).Success
+}